# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: 受験マップ API
  version: "1"
servers:
  - url: https://juken-map.com
security:
  - sessionCookie: []
paths:
...
  /api/admin/faculties:
    post:
      operationId: createFaculty
      summary: 学部を作る。同じ大学に同じ名前の学部があれば作らない
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFacultyInput" }
      responses:
        "201":
          description: 作った
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFacultySnapshot" }
        "400": { $ref: "#/components/responses/AdminFacultyBadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/AdminMasterConflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/admin/faculties/{id}:
    patch:
      operationId: updateFaculty
      summary: 学部を書き換える（名前・受験日・タグ。タグは送ったものに置き換える）
      parameters:
        - $ref: "#/components/parameters/AdminMasterID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FacultyInput" }
      responses:
        "200":
          description: 書き換えた後の学部
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFacultySnapshot" }
        "400": { $ref: "#/components/responses/AdminFacultyBadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/AdminMasterConflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteFaculty
      summary: 学部を消す。志望校に1件でも使われていれば消さない
      parameters:
        - $ref: "#/components/parameters/AdminMasterID"
      responses:
        "204": { description: 消した }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/AdminMasterConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-jm_session
      description: ログイン（Go の /api/auth/*、JUK-115）が発行するセッション Cookie。中身の無い 256 ビットの乱数のトークン
    cronBearer:
      type: http
      scheme: bearer
      description: DAILY_NOTIFICATION_SECRET
    simBearer:
      type: http
      scheme: bearer
      description: SIMULATION_SECRET（cron とは別のトークン）
  parameters:
...
    AdminMasterID:
      name: id
      in: path
      required: true
      description: >-
        マスターの行の ID。数字だけの正の整数（15桁まで）。形が違えば 400（ID と違い、ValidationError の形で
        code は invalid_format）
      schema: { type: string, pattern: "^[1-9][0-9]{0,14}$" }
...
  responses:
    AdminForbidden:
      description: >-
        管理者でない（{"error":"Forbidden"}）、2段階認証を通していない（code が TWO_FACTOR_REQUIRED）、
        利用を停止されている、別のサイトからの書き込み
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    AdminMasterConflict:
      description: >-
        名前・ISBN が重なっている、または使われていて消せない（文言に件数が入る。大学・学部は志望校、参考書は利用者の参考書）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AdminFacultyBadRequest:
      description: >-
        入力チェックで弾いた（ValidationError の形：文言・コード・項目）か、存在しないタグが含まれている
        （文言だけ）。どちらも error は必ずある
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: 入力チェックで弾いた（文言・コード・項目）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    Unauthorized:
      description: ログインしていない（またはトークンが違う）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: 見つからない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: 本文が上限（ふだんは 1MiB、CSP の報告は 16KB）を超えた
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
    UnsupportedMediaType:
      description: 受け付けない Content-Type（受けるのは JSON・text/plain・CSP の報告の2種類。本文が無ければ無しでよい）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
    InternalError:
      description: 想定外の失敗。原因はログにだけ残す
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
  schemas:
...
    AdminFacultySnapshot:
      description: 学部の作成・書き換えの応答（監査ログの「前・後」と同じ形）
      type: object
      required: [id, universityId, name, examDate, tagIds]
      properties:
        id: { type: integer, format: int64 }
        universityId: { type: integer, format: int64 }
        name: { type: string }
        examDate:
          description: YYYY-MM-DD
          type: string
          format: date
          x-go-type: string
        tagIds:
          type: array
          items: { type: integer, format: int64 }
...
    FacultyInput:
      description: >-
        学部の書き換え（Zod の facultyInputSchema）。受験日は YYYY-MM-DD で、JavaScript の new Date と同じく
        日は 31 まで受け付けて繰り上げる（2027-02-30 は 3月2日）。タグは重ねられない
      type: object
      required: [name, examDate, tagIds]
      properties:
        name: { type: string, minLength: 1, maxLength: 100, description: 前後の空白を削ってから数える }
        examDate: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
        tagIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: integer, format: int64, minimum: 1 }
    CreateFacultyInput:
      description: 学部の作成（Zod の createFacultySchema）。FacultyInput に大学を足したもの
      type: object
      required: [name, examDate, tagIds, universityId]
      properties:
        name: { type: string, minLength: 1, maxLength: 100, description: 前後の空白を削ってから数える }
        examDate: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
        tagIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: integer, format: int64, minimum: 1 }
        universityId: { type: integer, format: int64, minimum: 1 }
...
    Error:
      description: ルートが自分で断るときのエラー。error は画面にそのまま出す文言
      type: object
      required: [error]
      properties:
        error: { type: string }
    ValidationError:
      description: 入力チェックで弾いたときのエラー（Node の routes/validation-error.ts、Go の internal/httpx/response.go）
      type: object
      required: [error, code, field]
      properties:
        error:
          type: string
          description: 画面にそのまま出す文言
        code:
          type: string
          description: >-
            種類。自分で書いた規則は名前（range_end_before_start など）、
            組み込みのチェックは Zod の code（too_small・invalid_type・invalid_format など）
        field:
          type: string
          nullable: true
          description: どの項目か（"rangeEnd"・"items.0.unit"）。項目に結びつかなければ null
    ServerError:
      description: 想定外の失敗・混雑・存在しないパス
      type: object
      required: [error, code, reqId]
      properties:
        error: { type: string }
        code:
          type: string
          # FST_ERR_CTP_* は Fastify の code をそのまま使っている（Node の error-handling.ts）
          # FST_ERR_CTP_INVALID_CONTENT_LENGTH は本文が UTF-8 として不正で、読んだ長さが Content-Length と合わないとき（400）
          enum:
            - INTERNAL_ERROR
            - OVERLOADED
            - NOT_FOUND
            - FST_ERR_CTP_BODY_TOO_LARGE
            - FST_ERR_CTP_INVALID_MEDIA_TYPE
            - FST_ERR_CTP_INVALID_CONTENT_LENGTH
            # 利用者単位の回数制限を超えた（429、Go の internal/httpx/user_rate_limit.go）
            - TOO_MANY_REQUESTS
          x-enum-varnames:
            - ServerErrorCodeInternal
            - ServerErrorCodeOverloaded
            - ServerErrorCodeNotFound
            - ServerErrorCodeBodyTooLarge
            - ServerErrorCodeInvalidMediaType
            - ServerErrorCodeInvalidContentLength
            - ServerErrorCodeTooManyRequests
        reqId: { type: string }
...